Moved XmlEditerUtils to shared FXBEditerUtils. Removed xrmtb.XrmToolBox.Controls NO ILMerge anymore!
