+++ FXBEditorUtils/IComboBoxItem.cs
@@ -1,0 +1,7 @@
+﻿namespace Cinteros.Xrm.XmlEditorUtils
+{
+    public interface IComboBoxItem
+    {
+        string GetValue();
+    }
+}
