+++ FetchXmlBuilder/packages.config
@@ -2,7 +2,6 @@
 <packages>
   <package id="DockPanelSuite" version="3.0.6" targetFramework="net472" />
   <package id="DockPanelSuite.ThemeVS2015" version="3.0.6" targetFramework="net472" />
-  <package id="ILMerge" version="3.0.41" targetFramework="net472" />
   <package id="jacobslusser.ScintillaNET" version="3.6.3" targetFramework="net472" />
   <package id="Microsoft.CrmSdk.CoreAssemblies" version="9.0.2.32" targetFramework="net472" />
   <package id="Microsoft.CrmSdk.Deployment" version="9.0.2.32" targetFramework="net472" />
@@ -19,7 +18,6 @@
   <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
   <package id="NuGet.Common" version="5.9.1" targetFramework="net472" />
   <package id="NuGet.Configuration" version="5.9.1" targetFramework="net472" />
-  <package id="NuGet.Core" version="2.14.0" targetFramework="net452" />
   <package id="NuGet.Frameworks" version="5.9.1" targetFramework="net472" />
   <package id="NuGet.Packaging" version="5.9.1" targetFramework="net472" />
   <package id="NuGet.Versioning" version="5.9.1" targetFramework="net472" />
@@ -31,6 +29,5 @@
   <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net472" />
   <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net472" />
   <package id="System.Security.Cryptography.X509Certificates" version="4.3.2" targetFramework="net472" />
-  <package id="xrmtb.XrmToolBox.Controls" version="2.2.10" targetFramework="net472" />
   <package id="XrmToolBoxPackage" version="1.2021.5.47" targetFramework="net472" />
 </packages>
