--- XmlEditorUtils/IComboBoxItem.cs
@@ -1,7 +1,0 @@
-﻿namespace Cinteros.Xrm.XmlEditorUtils
-{
-    public interface IComboBoxItem
-    {
-        string GetValue();
-    }
-}
