--- XmlEditorUtils/IDefinitionSavable.cs
@@ -1,11 +1,0 @@
-﻿// PROJECT : MsCrmTools.SiteMapEditor
-// This project was developed by Tanguy Touzard
-// CODEPLEX: http://xrmtoolbox.codeplex.com
-// BLOG: http://mscrmtools.blogspot.com
-namespace Cinteros.Xrm.XmlEditorUtils
-{
-    public interface IDefinitionSavable
-    {
-        bool Save(bool keyPress);
-    }
-}
