--- XmlEditorUtils/packages.config
@@ -1,24 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="jacobslusser.ScintillaNET" version="3.6.3" targetFramework="net472" />
-  <package id="Microsoft.CrmSdk.CoreAssemblies" version="9.0.2.32" targetFramework="net472" />
-  <package id="Microsoft.CrmSdk.Deployment" version="9.0.2.32" targetFramework="net472" />
-  <package id="Microsoft.CrmSdk.Workflow" version="9.0.2.32" targetFramework="net472" />
-  <package id="Microsoft.CrmSdk.XrmTooling.CoreAssembly" version="9.1.0.68" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel" version="7.0.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="5.2.9" targetFramework="net472" />
-  <package id="Microsoft.NETCore.Platforms" version="1.1.1" targetFramework="net472" />
-  <package id="Microsoft.NETCore.Targets" version="1.1.3" targetFramework="net472" />
-  <package id="Microsoft.Rest.ClientRuntime" version="2.3.22" targetFramework="net472" />
-  <package id="Microsoft.Web.Xdt" version="3.1.0" targetFramework="net472" />
-  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
-  <package id="NuGet.Core" version="2.14.0" targetFramework="net452" />
-  <package id="System.IO" version="4.3.0" targetFramework="net472" />
-  <package id="System.Net.Http" version="4.3.4" targetFramework="net472" />
-  <package id="System.Private.Uri" version="4.3.2" targetFramework="net472" />
-  <package id="System.Runtime" version="4.3.1" targetFramework="net472" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.3.1" targetFramework="net472" />
-  <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.3.2" targetFramework="net472" />
-</packages>
